requirements.txt
# Core dependencies
litellm=1.82.6
python-dotenv>=0.21.0
google-generativeai>=0.8.0
